requirements.txt
pytorch==2.0.0
pytorch-cuda==11.7
torchvision==0.15.0
faiss-gpu==1.7.2
tqdm==4.66.1
pillow==10.0.1
scikit-learn==1.3.0
numpy==1.20.1
requests==2.31.0
einops==0.6.1
